+++ src/clj/clj_wamp/server.clj
@@ -40,7 +40,7 @@
 (def ^:const DESC-WAMP-ERROR-NOAUTH   "unauthorized")
 (def ^:const URI-WAMP-ERROR-NOAUTH    (str URI-WAMP-ERROR "unauthorized"))
 
-(def project-version "clj-wamp/1.0.0")
+(def project-version "clj-wamp/1.1.0-SNAPSHOT")
 
 (def max-sess-id (atom 0))